Abstracted default banner procedures.
