+++ lib/crawler/lib/schools/everyschool.js
@@ -48,6 +48,17 @@ var everyschool = module.exports = {
 			}
 			return this
 		}
+	, redefine: function(prop){
+			this[prop] = function (setTo) {
+				var k = '_' + prop;
+				if (arguments.length) {
+						this[k] = setTo
+						this[prop] = setTo;
+					return this
+				}
+			}
+			return this;
+		}
 	, mongoose: require('mongoose')
 	, url: 0
 	, paths: {}
@@ -120,14 +131,14 @@ var everyschool = module.exports = {
 			submodule.paths = {}
 			submodule.name = ''
 			
-      
-      var steps = this._steps
-        , newSteps = submodule._steps = {};
-      for (var stepName in steps) {
-        newSteps[stepName] = steps[stepName].clone(stepName, submodule);
-      }
+			
+			var steps = this._steps
+				, newSteps = submodule._steps = {};
+			for (var stepName in steps) {
+				newSteps[stepName] = clone(steps[stepName]);
+			}
 
-      return submodule;
+			return submodule;
 		}
